+++ Paper/JOSSpaper.bib
@@ -1,0 +1,6 @@
+@unpublished{LagCMAME,
+  title={A Lagrangian Formulation for Constrained Multibody Dynamics in Tensegrity Systems},
+  author={S.C. Hsu, V. Tadiparthi, and R. Bhattacharya},
+  note={Manuscript submitted for publication},
+  year={2018},
+  }
