### 2018-10-08 #1000 @arfon
> In my submission, I suggested that @labarba be assigned as an editor, due to her experience with fluid dynamics. Was there a problem with that?

No, that should be OK I think. @labarba - would you be able to handle this submission for JOSS?
