### 2018-10-08 #1000 @labarba
I'm a little worried that this software lists as a dependency the Matlab Optimization Toolbox, which is an additional thousand bucks over Matlab. How am I going to find reviewers for such a thing?
