### 2018-10-09 #1000 @ctdegroot
I have access to academic licenses for Matlab and the Optimization Toolbox. I would be able to do the review.
