### 2018-10-09 #1000 @moorepants
I'm sorry I do not have access to Matlab. If it runs in Octave I can review.
