### 2018-10-21 #1000 @labarba
@vaishnavtv : Your only reference in the paper is a manuscript that has been submitted but is not available without contacting the authors. Would you consider depositing your manuscript in a pre-print server? You can use arXiv, [engrXiv](https://engrxiv.org), or other. A majority of [journals accept preprints](https://en.wikipedia.org/wiki/List_of_academic_journals_by_preprint_policy) these days.
