### 2018-10-21 #1000 @labarba
@Kevin-Mattheus-Moerman : Would you be able to review this submission for JOSS? It uses the Matlab Optimization Toolbox.
