### 2018-10-22 #1000 @Kevin-Mattheus-Moerman
Perhaps folks at https://github.com/NASA-Tensegrity-Robotics-Toolkit e.g. @apsabelhaus (who contributed to https://github.com/apsabelhaus/tInvKin) would be interested in reviewing this?
