### 2018-10-22 #1000 @Kevin-Mattheus-Moerman
My colleague at MIT might be interested in this too. @cybanical fancy testing some code to study dynamics of tensegrity structures? (also I recommend that you upgrade your github profile with an ORCID link, your real name, and some affiliation information etc).
