### 2018-10-22 uqLab/stedy#1 @Kevin-Mattheus-Moerman
README issues
1. Please link to contributing guidelines and code of conduct in your README.
2. For the startup instructions it seems one flow is 1, 2, 3, 6, if no error occurs and all steps if an error does occur. This does not seem so clear. Perhaps rename step 6 to 4 and add a note to step 3 to say that if an error occurs please do steps 3b and 3c below. Or change this in a similar way to clarify the two processes.
3. Please consider splitting the startup section into two sections, e.g. 1) Installation and setup instructions, 2) Getting started.
