### 2018-10-22 uqLab/stedy#2 @Kevin-Mattheus-Moerman
Example TemplateForUser_DBar.m does not exist
The [user guide](https://github.com/uqLab/stedy/blob/master/Docs/UserGuide.md) refers to `TemplateForUser_DBar.m`, however this file does not exist. Perhaps this should be `Template_DBar.m`?
