### 2018-10-22 uqLab/stedy#3 @Kevin-Mattheus-Moerman
Comments about / errors in example: Template_DBar
About `Template_DBar.m`

1. When I run this example I receive a question on the command prompt `Enter output time-step:`, please can you add a suggested sensible time-step? E.g. rephrase it as `Enter output time-step (e.g. 1):`
2. After entering 1 as the output time-step I get the following error message:

 ```none
Error using animateTenseg (line 29)
The specified profile is not valid.

Error in Template_DBar (line 147)
animateTenseg(t,y,tData,time_loc,'DBar_Animation.mp4','MPEG-4',frameRate,AZ, EL,axLims);
```

I am using a 64-bit Ubuntu 18.04 machine with MATLAB R2018b.
