### 2018-10-22 uqLab/stedy#3 @Kevin-Mattheus-Moerman
I get the same error for other examples
