### 2018-10-22 uqLab/stedy#3 @Kevin-Mattheus-Moerman
Side note, if you want to avoid issues like this you could animate it in MATLAB directly (e.g. see the [`anim8`](https://github.com/gibbonCode/GIBBON/blob/master/docs/HELP_anim8.m) function in the [GIBBON](https://github.com/gibbonCode/GIBBON) toolbox I maintain).
