### 2018-10-22 uqLab/stedy#3 @vaishnavtv
Thank you @Kevin-Mattheus-Moerman, I'll make the suggested change to the output time-step prompt. 0.01 is usually chosen as the output time-step to ensure higher accuracy.
It is generally recommended that the 'Refine' option be turned on in ODE options to avoid errors due to numerical extrapolation, but I see that the Template file itself doesn't have it included in the options. I'll make the change there and also edit the User Guide to reflect the prompt change.

I'll also make the necessary changes as suggested about the VideoWriter format. Thank you for the GIBBON recommendation. I'll take a look.
