### 2018-10-22 uqLab/stedy#1 @vaishnavtv
Will incorporate the suggested changes, thank you.
