### 2018-10-22 911e5c5
Function to generate SimTime, tInt
