+++ Docs/UserGuide.md
@@ -36,8 +36,8 @@ The dynamics of tensegrity systems is governed by multi-body dynamics, given by 
 A common example of a tensegrity system is that of a D-Bar.
 ![DBar_Configuration](Figs/DBar.jpg)
 
-A DBar is a tensegrity system of 4 bars (in black) and 2 strings (in red) connected as shown above. The initial nodal configuration, the connectivity matrices, the material and geometric properties as well as the simulation environment properties are described below in further detail.  
-The user is recommended to go through TemplateForUser_DBar.m in the examples folder to view the entire source file for the following inputs.
+A DBar is a tensegrity system of 4 bars (in black) and 2 strings (in red) connected as shown above. The initial nodal configuration, the connectivity matrices, the material and geometric properties as well as the simulation environment properties are described below in further detail.
+The user is recommended to go through Template_DBar.m in the examples folder to view the entire source file for the following inputs.
 
 ## Inputs Required from the User:
 All parameters are to be declared in SI units.
@@ -52,8 +52,8 @@ All parameters are to be declared in SI units.
     %      N1y N2y N3y ..;
     %      N1z N2z N3z ..;];
     ```
-2. **C**: A matrix of size *(m x nNodes)* <!--![equation](https://latex.codecogs.com/gif.latex?\in&space;\mathbb{R}^{m&space;\times&space;n})--><!--($\in \mathbb{R}^{m \times n}$)--> defines the connectivity matrix describing the network locations of *m* <!--![equation](https://latex.codecogs.com/gif.latex?m)--><!-- $m$ --> bars and strings.  
-Specifically, if the kth <!--![equation](https://latex.codecogs.com/gif.latex?k^{th})--><!--$k^{th}$--> member is defined by connecting nodes *n(i)* <!--![equation](https://latex.codecogs.com/gif.latex?n_i)--><!--$n_i$--> and *n(j)*, <!--![equation](https://latex.codecogs.com/gif.latex?n_j)--><!--$n_j$-->   
+2. **C**: A matrix of size *(m x nNodes)* <!--![equation](https://latex.codecogs.com/gif.latex?\in&space;\mathbb{R}^{m&space;\times&space;n})--><!--($\in \mathbb{R}^{m \times n}$)--> defines the connectivity matrix describing the network locations of *m* <!--![equation](https://latex.codecogs.com/gif.latex?m)--><!-- $m$ --> bars and strings.
+Specifically, if the kth <!--![equation](https://latex.codecogs.com/gif.latex?k^{th})--><!--$k^{th}$--> member is defined by connecting nodes *n(i)* <!--![equation](https://latex.codecogs.com/gif.latex?n_i)--><!--$n_i$--> and *n(j)*, <!--![equation](https://latex.codecogs.com/gif.latex?n_j)--><!--$n_j$-->
  *C(k,i) = -1, C(k,j) = 1*. All other elements in the kth row should be zero.
 
     There are two ways the user can characterize these matrices:
@@ -111,7 +111,7 @@ Strings are modeled as springs but unlike springs, they can only exert tensile f
       strings.r = [0.01 0.05 0.01 0.05];
       ```
 
-      Rest length here is in percentage,i.e., 0.7 implies 70%. The rest lengths of the strings are 70% of the current length of the strings.  
+      Rest length here is in percentage,i.e., 0.7 implies 70%. The rest lengths of the strings are 70% of the current length of the strings.
       **NOTE**: The rest lengths of the strings are a measure of the pre-stress in the structure. The lower the rest lengths, the more pre-stressed a structure is.
 5. **bars**: MATLAB structure containing fields to define material properties of the bars present in the tensegrity structure.
[…]
