### 2018-10-22 b78df70
Removed simTime generation from tensegSim, added PC OS check in animateTenseg
