### 2018-10-22 9e1f030
Updated example files to reflect changes in tensegSim and animateTenseg
