+++ Docs/README.md
@@ -64,6 +64,7 @@ The software is aimed at researchers familiar with tensegrity structures, but ca
       ```matlab
       function varargout = ode45m(ode,tspan,y0,options,varargin)
       ```
+      
 4. To avoid having to run setup.m everytime you run MATLAB, go to *Set Path* in the Home tab and add the corresponding location of *stedy/Main* to MATLAB's path.
 
 
