### 2018-10-22 uqLab/stedy#1 @vaishnavtv
Fixed this, let me know if you'd like any further changes to be made.
