### 2018-10-22 uqLab/stedy#3 @vaishnavtv
Fixed. Added OS check to animateTenseg.m and updated prompt message for output time-step.
