### 2018-10-22 #1000 @apsabelhaus
Dear all - very nice to meet you!

I'm happy to be a reviewer, though I'm not sure what this entails, I'm new to JOSS. I have the MATLAB Optimization Toolbox.

On a different note, the field of dynamics derivations for tensegrity structures is actually very extensive, and it would be very good to include many other knowledgeable experts in the area. There are well-known expressions for the dynamics of these structures that could be verified against the authors' code. In fact, the authors have a colleague at Texas A&M who is a forerunner in the area - Robert Skelton - who can refer good reviewers for this code. Other significant authors in the field include Cornel Sultan, Josep Mirats-Tur, and their students.

Of particular note would be other contributors to NTRTsim (https://github.com/NASA-Tensegrity-Robotics-Toolkit) such as Brian Mirletz @brtietz and the current maintainer Jonathan Bruce @JEB12345. Also, Jeffrey Friesen at @Jfriesen222 has already derived the dynamics for a large set of tensegrity robots.

Please let me know how you might like from me.
