### 2018-10-22 #1042 @whedon
[REVIEW]: STEDY - Software for TEnsegrity DYnamics
**Submitting author:** @vaishnavtv (<a href="http://orcid.org/0000-0001-9897-0600">Venkata Vaishnav Tadiparthi</a>)
**Repository:** <a href="https://github.com/uqLab/stedy" target ="_blank">https://github.com/uqLab/stedy</a>
**Version:** v0.1.0
