### 2018-10-23 #1042 @Kevin-Mattheus-Moerman
@labarba over at https://github.com/openjournals/joss-reviews/issues/1000# @apsabelhaus agreed to be reviewer, perhaps you could add him here.
