### 2018-10-23 #1042 @Kevin-Mattheus-Moerman
@apsabelhaus let me introduce you to JOSS and reviewing for JOSS. [JOSS](http://joss.theoj.org/) is a free, open access, and open source journal. Reviews take place here, i.e. in Github issues. As you can see the top of this issue has checklists for the reviewers which will guide them through the review process. A set of checklists will be created for you if you join as reviewer. As described in our [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) the review focuses on [the software](https://github.com/uqLab/stedy) and a [short paper](https://github.com/openjournals/joss-papers/blob/joss.01042/joss.01042/10.21105.joss.01042.pdf). Reviewers can comment on this work here in this issue and are encouraged to created dedicated issues on the software repository, and link to them here, for larger items. Let me know if you have more questions.
