### 2018-10-23 #1042 @Kevin-Mattheus-Moerman
@labarba so far @vaishnavtv has worked on has, and resolved, these issues which I've posted on the software repository:

- [ ] https://github.com/uqLab/stedy/issues/1 README issues
- [ ] https://github.com/uqLab/stedy/issues/2 Example TemplateForUser_DBar.m does not exist
- [ ] https://github.com/uqLab/stedy/issues/3 Comments about / errors in example: Template_DBar
