### 2018-10-23 #1042 @labarba
@Kevin-Mattheus-Moerman : I forgot how to give @apsabelhaus the permissions to edit the reviewer checklist or send him the appropriate invitation. Do you remember? I went to the Teams section of the `open journals` org, and couldn't immediately identify a team to add him. Help…
