### 2018-10-24 #1042 @Kevin-Mattheus-Moerman
Then this invitation should be accepted: https://github.com/openjournals/joss-reviews/invitations
