### 2018-10-24 #1042 @labarba
The `@whedon add` commands sets things up _only_ if used in the Pre-Review issue, _before_ starting the review. Once we're here, adding a reviewer is a manual process.
