### 2018-10-31 #1000 @cybanical
@Kevin-Mattheus-Moerman <https://github.com/Kevin-Mattheus-Moerman> This
looks super interesting, but it looks like I'm too slammed right now. I've
got two other papers to review right now for ICRA.
*_____________________________________*
*MATTHEW CARNEY | Research Assistant and PhD Candidate **| *Mechatronics
Design Engineer
MIT Media Lab | Biomechatronics
75 Amherst St. | E14-274G | Cambridge, MA 02142 USA
biomech.media.mit.edu
<https://www.media.mit.edu/groups/biomechatronics/overview/> |
matthematic.com | IG:@matt.thematic
<https://www.instagram.com/matt.thematic/>

On Tue, Oct 23, 2018 at 3:45 AM Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @apsabelhaus <https://github.com/apsabelhaus> thanks for those
> suggestions!
>
