### 2018-11-01 #1042 @vaishnavtv
You can find the pre-print of the cited reference [here](https://www.researchgate.net/publication/328676032_A_Lagrangian_Formulation_for_Constrained_Multibody_Dynamics_in_Tensegrity_Systems).

I hope the review is going smoothly. Please let us know if anything else is required.
