### 2018-11-15 #1042 @ctdegroot
@vaishnavtv The paper contains insufficient references and the one reference that is included is not formatted properly (it appears twice). Please include at least a couple more references. For example, you mention some similar software called "Simscape Multibody". You should reference that.  The first paragraph needs a reference or two also.
