+++ Paper/JOSSpaper.bib
@@ -2,5 +2,57 @@
   title={A Lagrangian Formulation for Constrained Multibody Dynamics in Tensegrity Systems},
   author={S.C. Hsu, V. Tadiparthi, and R. Bhattacharya},
   note={Manuscript submitted for publication},
+  DOI = {10.13140/RG.2.2.34594.89284}
   year={2018},
+  }
+
+@book{tenseBook,
+    title={Tensegrity systems},
+    author={Skelton, Robert E and de Oliveira, Mauricio C},
+    volume={1},
+    year={2009},
+    publisher={Springer}
+  }
+
+  @book{matlab,
+  year = {2016},
+  author = {MATLAB},
+  title = {version 9.0 (R2016a)},
+  publisher = {The MathWorks Inc.},
+  address = {Natick, Massachusetts}
+  }
+
+  @Article{ConstraintViolation,
+  author="Marques, Filipe
+  and Souto, Ant{\'o}nio P.
+  and Flores, Paulo",
+  title="On the constraints violation in forward dynamics of multibody systems",
+  journal="Multibody System Dynamics",
+  year="2017",
+  month="Apr",
+  day="01",
+  volume="39",
+  number="4",
+  pages="385--419",
+  issn="1573-272X",
+  doi="10.1007/s11044-016-9530-y",
+  url="https://doi.org/10.1007/s11044-016-9530-y"
+  }
+
+  @article{ingber1998architecture,
+    title={The architecture of life},
+    author={Ingber, Donald E},
+    journal={Scientific American},
+    volume={278},
+    number={1},
+    pages={48--57},
+    year={1998},
+    publisher={JSTOR}
+  }
+
+  @book{scarr2014biotensegrity,
+    title={Biotensegrity},
+    author={Scarr, Graham},
+    year={2014},
+    publisher={Handspring Publishing, United Kingdom}
   }
