+++ Paper/JOSSpaper.md
@@ -18,13 +18,13 @@ authors:
 
   - name: Raktim Bhattacharya
     orcid: 0000-0002-8218-2631
-    affiliation: 2
+    affiliation: 1
 affiliations:
- - name: Graduate Research Assistant, Texas A&M University
+ - name: Department of Aerospace Engineering, Texas A&M University, College Station, Texas, USA
    index: 1
 
- - name: Associate Professor, Texas A&M University
-   index: 2
+ # - name: Associate Professor, Texas A&M University
+ #   index: 2
 
 date: 13 September 2018
 bibliography: JOSSpaper.bib
@@ -32,24 +32,24 @@ bibliography: JOSSpaper.bib
 
 # Summary
 
-A tensegrity system is an arrangement of axially-loaded elements (no element bends, even though the overall structure bends), that we loosely characterize as a network of bars and cables. The bars take compressive axial loads and the cables handle tensile loads. Since failure due to axial stresses happens at higher loads than at bending, a tensegrity structure has a higher strength-to-weight ratio. The famous architect Buckminster Fuller, in the 60's, coined the term tensegrity, combining the words tensile and integrity. Since then, tensegrity principles have found applications in diverse domains like architecture, biological modeling as well as civil engineering design. Tensegrity structures, through use of pre-stresses in the bars and cables, can also achieve controlled stiffness in the structure, which makes it attractive in applications such as soft-robotics and prosthetics. In essence, tensegrity principles can be applied in the design of any structure where mass is premium, a high strength-to-weight ratio is critical, and structural stiffness needs to be tailored in both space and time. These include several applications from various engineering sectors such as aerospace (morphing airframes), energy (wind turbine blades, off-shore structures) as well as biomedical engineering (stents, minimally invasive surgical tools) and many more. Clearly, a framework is required that can efficiently model the dynamics of tensegrity structures directly from the topology of bars and cables.
+A tensegrity system is an arrangement of axially-loaded elements (no element bends, even though the overall structure bends), that we loosely characterize as a network of bars and cables. The bars take compressive axial loads and the cables handle tensile loads. Since failure due to axial stresses happens at higher loads than at bending, a tensegrity structure has a higher strength-to-weight ratio. The famous architect Buckminster Fuller, in the 60's, coined the term tensegrity, combining the words tensile and integrity. Since then, tensegrity principles have found applications in diverse domains like architecture [@ingber1998architecture], biological modeling [@scarr2014biotensegrity] as well as civil engineering design [@tenseBook]. Tensegrity structures, through use of pre-stresses in the bars and cables, can also achieve controlled stiffness in the structure, which makes it attractive in applications such as soft-robotics and prosthetics. In essence, tensegrity principles can be applied in the design of any structure where mass is premium, a high strength-to-weight ratio is critical, and structural stiffness needs to be tailored in both space and time. These include several applications from various engineering sectors such as aerospace (morphing airframes), energy (wind turbine blades, off-shore structures) as well as biomedical engineering (stents, minimally invasive surgical tools) and many more. Clearly, a framework is required that can efficiently model the dynamics of tensegrity structures directly from the topology of bars and cables.
 
[…]
