+++ Paper/JOSSpaper.bib
@@ -2,7 +2,7 @@
   title={A Lagrangian Formulation for Constrained Multibody Dynamics in Tensegrity Systems},
   author={S.C. Hsu, V. Tadiparthi, and R. Bhattacharya},
   note={Manuscript submitted for publication},
-  DOI = {10.13140/RG.2.2.34594.89284}
+  DOI = {10.13140/RG.2.2.34594.89284},
   year={2018},
   }
 
