### 2018-11-15 #1042 @vaishnavtv
@ctdegroot I have made the recommended changes to the paper - affiliations and references have been added. We were under the impression that our manuscript submitted for publication would suffice as it exhaustively covers all information pertinent to this submission as well. Thank you for the clarification.

For the automated tests, would the files provided in the Examples folder be sufficient?
