### 2018-11-16 #1042 @ctdegroot
@vaishnavtv The paper looks good now.

I don't think that the files in the Examples folder are sufficient as test cases because they do not assert what result is expected. Let's imagine that I am a user that makes some changes/extensions to the code. How to I check that I haven't broken anything that already existed?

From the JOSS review criteria:

Good: An automated test suite hooked up to an external service such as Travis-CI or similar
OK: Documented manual steps that can be followed to check the expected functionality of the software (e.g. a sample input file to assert behaviour)
Bad (not acceptable): No way for you the reviewer to check whether the software works

To get the tests into the 'OK' category you will need to provide some reference results that can be checked (preferably automatically) with the code outputs.

One other note, is that you should remove all of the ".DS_Store" files that are in some of the example folders. Normally you can put this in your .gitignore file to prevent them from being tracked in the future.
