### 2018-11-16 6ba9821
tensegGenMat updated to allow cases with no bars or no strings, Test Case for pendulum added
