### 2018-11-16 1c9ed47
Added testConstraint to check for constraint violations
