### 2018-11-16 #1042 @vaishnavtv
@ctdegroot
As recommended, I have added a Tests folder to the repository. The user can simply run TestPendulum.m to check the results of their changes to the code. I have also modified the Contributing Guidelines document to reflect this update.

I have removed all the .DS_Store files as well.
