### 2018-11-19 #1042 @ctdegroot
@labarba I'm satisfied with the changes so far. Still need to finish checking on the functionality of the code; hopefully soon.
