### 2018-11-20 uqLab/stedy#6 @Kevin-Mattheus-Moerman
Community guidelines
Can you please work on adding community guidelines by adding a CONTRIBUTING.md file (and by linking to it in your readme)? See also this link for more information: https://help.github.com/articles/setting-guidelines-for-repository-contributors/

I also recommend adding a CODE_OF_CONDUCT.md file (to which you can link in the contributing file).

You can get a template for the code of conduct file from here: https://www.contributor-covenant.org/version/1/4/code-of-conduct.html

You can also check out how my personal project uses a contibuting and coc file:
https://github.com/gibbonCode/GIBBON

Let me know if you have questions.
