### 2018-11-20 #1042 @Kevin-Mattheus-Moerman
@vaishnavtv can you please work on adding basic community guidelines? See this issue:
- [ ] [Community Guidelines](https://github.com/uqLab/stedy/issues/6)
