### 2018-11-20 #1042 @Kevin-Mattheus-Moerman
@labarba thanks for the reminder. Only one box now remains unticked and it relates to community guidelines. I've opened an issue on this an am awaiting feedback from the author.
