### 2018-11-20 uqLab/stedy#6 @vaishnavtv
@Kevin-Mattheus-Moerman  the [readme] already consists of a Documents subsection containing links to the Contributing guidelines and the Code of Conduct along with the User Guide. Did you want me to link them in a different manner?

[readme]: https://github.com/uqLab/stedy/blob/master/Docs/README.md
