### 2018-11-20 #1042 @apsabelhaus
@Kevin-Mattheus-Moerman it doesn't seem like I'm able to fill out the checklist. Following the link to https://github.com/openjournals/joss-reviews/invitations seems to imply I haven't been formally sent an invitation. Could you do so / am I missing something? I'll proceed with evaluating the software in the meantime! Thanks again for everyone's patience, I'm more used to traditional academic peer-review deadlines.
