### 2018-11-20 #1042 @labarba
@arfon Can you help us? I added @apsabelhaus as reviewer manually here, after the review started, so we don't have the benefit of our bot `whedon` setting the permissions right. I don't remember how to give the permissions to the reviewer, in this case. I looked at the teams in the org, but it's not clear to me: a `joss-reviewers` team exists, but it has only two members.
