### 2018-11-20 #1042 @arfon
@labarba - I've added @apsabelhaus as an external collaborator to this repository which will will allow them to update the checklist.  @apsabelhaus - the invite is available here: https://github.com/openjournals/joss-reviews/invitations
