### 2018-11-20 #1042 @apsabelhaus
Thank you @arfon and @labarba ! Started checking off the list.

My initial review is that this is a very nice piece of software, with a well-defined use case, filling a gap that the community needs. With some minor changes and clarifications, I would be ready to recommend for publication. @vaishnavtv , I've opened a handful of issues, mostly about documentation or technical questions. I've currently left the following boxes unticked for these reasons:

1) Functionality: one minor potential bug about the video.
2) Functionality documentation: some confusion about the API and technical assumptions.
4) References: just a bit more explanation needed about certain technical parts that weren't referred to in the JOSS paper or your preprint. I've provided some examples to refer to.

Thanks to the authors for a great contribution to the tensegrity systems community!
