+++ Main/plotMotion.m
@@ -2,7 +2,7 @@ function plotMotion(t,y,tData)
 % /* This Source Code Form is subject to the terms of the Mozilla Public
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
 % * file, You can obtain one at http://mozilla.org/MPL/2.0/. 
-
+% 
 % Function to plot trajectories of all nodes in the tensegrity structure
 % after solution has been obtained. 
 %
