+++ Main/ODE_helperFiles/odearguments.m
@@ -1,0 +1,158 @@
+function [neq, tspan, ntspan, next, t0, tfinal, tdir, y0, f0, args, odeFcn, ...
+          options, threshold, rtol, normcontrol, normy, hmax, htry, htspan, ...
+          dataType ] =   ...
+    odearguments(FcnHandlesUsed, solver, ode, tspan, y0, options, extras)
+%ODEARGUMENTS  Helper function that processes arguments for all ODE solvers.
+%
+%   See also ODE113, ODE15I, ODE15S, ODE23, ODE23S, ODE23T, ODE23TB, ODE45.
+
+%   Mike Karr, Jacek Kierzenka
+%   Copyright 1984-2011 The MathWorks, Inc.
+
+if strcmp(solver,'ode15i')
+  FcnHandlesUsed = true;   % no MATLAB v. 5 legacy for ODE15I
+end  
+
+if FcnHandlesUsed  % function handles used
+  if isempty(tspan) || isempty(y0) 
+    error(message('MATLAB:odearguments:TspanOrY0NotSupplied', solver));
+  end      
+  if length(tspan) < 2
+    error(message('MATLAB:odearguments:SizeTspan', solver));
+  end  
+  htspan = abs(tspan(2) - tspan(1));  
+  tspan = tspan(:);
+  ntspan = length(tspan);
+  t0 = tspan(1);  
+  next = 2;       % next entry in tspan
+  tfinal = tspan(end);     
+  args = extras;                 % use f(t,y,p1,p2...) 
+
+else  % ode-file used   (ignored when solver == ODE15I)
+  % Get default tspan and y0 from the function if none are specified.
+  if isempty(tspan) || isempty(y0) 
+    if exist(ode)==2 && ( nargout(ode)<3 && nargout(ode)~=-1 ) 
+      error(message('MATLAB:odearguments:NoDefaultParams', funstring( ode ), solver, funstring( ode )));      
+    end
+    [def_tspan,def_y0,def_options] = feval(ode,[],[],'init',extras{:});
+    if isempty(tspan)
+      tspan = def_tspan;
+    end
+    if isempty(y0)
+      y0 = def_y0;
+    end
+    options = odeset(def_options,options);
+  end  
+  tspan = tspan(:);
+  ntspan = length(tspan);
+  if ntspan == 1    % Integrate from 0 to tspan   
+    t0 = 0;          
+    next = 1;       % Next entry in tspan.
+  else              
+    t0 = tspan(1);  
+    next = 2;       % next entry in tspan
+  end
+  htspan = abs(tspan(next) - t0);
+  tfinal = tspan(end);   
+  
+  % The input arguments of f determine the args to use to evaluate f.
+  if (exist(ode)==2)
+    if (nargin(ode) == 2)           
+      args = {};                   % f(t,y)
+    else
+      args = [{''} extras];        % f(t,y,'',p1,p2...)
+    end
+  else  % MEX-files, etc.
+    try 
+      args = [{''} extras];        % try f(t,y,'',p1,p2...)     
+      feval(ode,tspan(1),y0(:),args{:});   
+    catch ME
+      args = {};                   % use f(t,y) only
+    end
+  end
+end
+
+y0 = y0(:);
+neq = length(y0);
+
+% Test that tspan is internally consistent.
+if t0 == tfinal
+  error(message('MATLAB:odearguments:TspanEndpointsNotDistinct'));
+end
+tdir = sign(tfinal - t0);
+if any( tdir*diff(tspan) <= 0 )
+  error(message('MATLAB:odearguments:TspanNotMonotonic'));
[…]
