+++ Main/tensegSimTime.m
@@ -2,7 +2,7 @@ function [simTime, tInt] = tensegSimTime(options,tEnd)
 % /* This Source Code Form is subject to the terms of the Mozilla Public
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
 % * file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
+% 
 % Function to generate simulation time and output interval time-step for
 % simulation and animation. 
 % 
