+++ Main/tensegSim.m
@@ -2,8 +2,7 @@ function [t,y] = tensegSim(x0,simTime,tData,options)
 % /* This Source Code Form is subject to the terms of the Mozilla Public
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
 % * file, You can obtain one at http://mozilla.org/MPL/2.0/. 
-
-
+% 
 % Function to simulate the tensegrity structure's dynamics in the
 % prescribed environment
 %
