+++ Main/ODE_helperFiles/bvpfunctions.m
@@ -1,269 +1,0 @@
-function [odefinal,bcfinal,jacfinal,bcjacfinal,Joptions,dBCoptions] = ...
-        bvpfunctions(solver_name,odefun,bcfun,options,...
-                     neqn,nparam,nregions,extras)
-%BVPFUNCTIONS  Helper function for processing functional arguments for BVP solvers.
-%
-%   See also BVP4C, BVP5C, BVPSET.
-
-%   Copyright 2007-2008 The MathWorks, Inc.
-
-    % handle multipoint BVP
-    if nargin < 7
-        nregions = 1;
-    end
-    
-    % handle extra arguments
-    if nargin < 8
-        extras = {};
-    end
-    
-    % Jacobian functions
-    jac = bvpget(options,'FJacobian',[]);
-    if ~( isempty(jac) || isa(jac,'function_handle') || ischar(jac) || isa(jac,'inline') || ...
-          (nparam == 0 && isnumeric(jac) || (nparam > 0 && iscell(jac)))) 
-        error(message('MATLAB:bvpfunctions:InvalidFJacobian'));  
-    end
-    bcjac = bvpget(options,'BCJacobian',[]);
-    if ~( isempty(bcjac) || isa(bcjac,'function_handle') || ischar(bcjac) || ...
-          iscell(bcjac) || isa(bcjac,'inline')) 
-        error(message('MATLAB:bvpfunctions:InvalidBCJacobian'));  
-    end            
-    
-    % make functions feval-free
-    ode = fcnchk(odefun);
-    bc  = fcnchk(bcfun);
-    if ischar(jac)
-        jac = str2func(jac);
-    end
-    if ischar(bcjac)
-        bcjac = str2func(bcjac);
-    end        
-    
-    odefinal = ode;
-    bcfinal = bc;
-    jacfinal = jac;
-    bcjacfinal = bcjac;    
-    
-    % incorporate known parameters (for BVP4C)
-    if ~isempty(extras) && strcmp(solver_name,'bvp4c')
-        if nparam == 0 
-            if nregions == 1
-                odefinal = @(x,y) ode(x,y,extras{:});
-                if isa(jac,'function_handle')
-                    jacfinal = @(x,y) jac(x,y,extras{:});
-                end
-            else
-                odefinal = @(x,y,region) ode(x,y,region,extras{:});
-                if isa(jac,'function_handle')
-                    jacfinal = @(x,y,region) jac(x,y,region,extras{:});
-                end
-            end
-            bcfinal = @(ya,yb) bc(ya,yb,extras{:});
-            if isa(bcjac,'function_handle')
-                bcjacfinal = @(ya,yb) bcjac(ya,yb,extras{:});
-            end
-        else
-            if nregions == 1
-                odefinal = @(x,y,p) ode(x,y,p,extras{:});
-                if isa(jac,'function_handle')
-                    jacfinal = @(x,y,p) jac(x,y,p,extras{:});
-                end
-            else
-                odefinal = @(x,y,region,p) ode(x,y,region,p,extras{:});
-                if isa(jac,'function_handle')
-                    jacfinal = @(x,y,region,p) jac(x,y,region,p,extras{:});
-                end
-            end                
-            bcfinal = @(ya,yb,p) bc(ya,yb,p,extras{:});
-            if isa(bcjac,'function_handle')
-                bcjacfinal = @(ya,yb,p) bcjac(ya,yb,p,extras{:});
-            end
-        end
-    end            
-    
-    % incorporate unknown parameters (for BVP5C)
-    if (nparam > 0) && strcmp(solver_name,'bvp5c')
-        if nregions == 1
-            odefinal = @odeParameters;
-            if ~isempty(jac) 
-                jacfinal = @jacParameters;
-            end
[…]
