+++ Examples/Arm/Arm.m
@@ -1,7 +1,7 @@
 % /* This Source Code Form is subject to the terms of the Mozilla Public
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
 % * file, You can obtain one at http://mozilla.org/MPL/2.0/. 
-
+% 
 % This example demonstrates a tensegrity-based robotic arm with 16 bars and
 % 18 strings. 
 %
