+++ Tests/testConstraint.m
@@ -2,7 +2,7 @@ function testConstraint(t,y,tData)
 % /* This Source Code Form is subject to the terms of the Mozilla Public
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
 % * file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
+% 
 % Function to test constraint violations.
 
 nn = numel(tData.N);
