+++ Tests/TestPendulum.m
@@ -1,7 +1,11 @@
 % /* This Source Code Form is subject to the terms of the Mozilla Public
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
 % * file, You can obtain one at http://mozilla.org/MPL/2.0/. 
-
+% 
+% This file provides a test case for the user to check their version of the
+% dynamics simulation with the results of a minimum-coordinates
+% realisation, conventionally accepted as the ground truth. 
+% 
 clc; clear;
 
 %% Define initial nodal coordinates and connectivity matrices
