### 2018-11-20 b1efb50
function help info
