### 2018-11-21 #1042 @ctdegroot
@vaishnavtv You need to revise your installation instructions. It is not clear that steps 3b and 3c are only required if there is an error. I had to read several times to make sense of it.
