### 2018-11-21 #1042 @ctdegroot
@vaishnavtv In the test that you've added, please do not make the user interactively enter data. If I'm trying to test the code as a new user, I don't know what I should enter. Is one test enough to verify the full functioning of the code?
