### 2018-11-21 #1042 @ctdegroot
@vaishnavtv Regarding the item "Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)": You have stated in your paper that "We have demonstrated superiority of the method in terms of accuracy over the commercially avail- able dynamics simulator, Simscape Multibody". There is a performance claim made. How do I confirm it?
