### 2018-11-21 #1042 @labarba
Thank you all for your careful work!
@ctdegroot We don't _require_ test coverage or unit tests vs. regression tests.

The JOSS Review [criteria on tests](https://joss.readthedocs.io/en/latest/review_criteria.html?highlight=test#tests) are:

> Authors are strongly encouraged to include an automated test suite covering the core functionality of their software.
> Good: An automated test suite hooked up to an external service such as Travis-CI or similar
> OK: Documented manual steps that can be followed to check the expected functionality of the software (e.g. a sample input file to assert behaviour)
> Bad (not acceptable): No way for you the reviewer to check whether the software works
