### 2018-11-21 #1042 @ctdegroot
@labarba Thank you for the clarification. In this case I still think it falls into the not acceptable category. I ran the one test that was provided and based on the output I don't know if the code is working or not. It runs and shows some plots, but is it working correctly? Not sure.
