### 2018-11-21 #1042 @vaishnavtv
@ctdegroot, when you run the test case, you should see the results of 3 checks (bar length, energy constraints and motion errors) in the command window. Let me know if that’s not being displayed.
