### 2018-11-21 #1042 @vaishnavtv
@ctdegroot , about the performance claim issue, we demonstrated superiority in terms of accuracy over Simscape multibody and attached plots in the [preprint]. Would you like us to attach the relevant files to the repository?
[preprint]: https://www.researchgate.net/publication/328676032_A_Lagrangian_Formulation_for_Constrained_Multibody_Dynamics_in_Tensegrity_Systems
