+++ Main/plot_configuration.m
@@ -57,12 +57,16 @@ figure(figOptions);hold on;
             plot3(NN(1,k),NN(2,k),NN(3,k),'ko','MarkerFaceColor','k')
         end
     end
-
+    quiver3(zeros(3,1),zeros(3,1),zeros(3,1),[1;0;0],[0;1;0],[0;0;1],'Color','g')
+    text(0.9,0,0,'x','fontsize',20,'color','b');
+    text(0,0.9,0,'y','fontsize',20,'color','b');
+    text(0,0,0.9,'z','fontsize',20,'color','b');
+    
 %     for i=1:tData.nNodes
 %         txt{i} = num2str(i);
 %         text(tData.N(1,i),tData.N(2,i),tData.N(3,i),txt{i},'fontsize',30);
 %     end
-    axis equal;
-    axis off;
+%     axis equal;
+%     axis off;
     view(AZ,EL);
 end
