### 2018-11-21 #1042 @ctdegroot
@vaishnavtv The preprint is reference in the paper right? That is good enough for me to verify the performance claim. For the test, what are you comparing to for a reference? Is it an exact solution? It is not clear to me from inspecting the code or running the case. There are no comments in "Pendulum_MinRealODE.m" so I don't know what it does.
