### 2018-11-21 #1042 @vaishnavtv
The test compares the results of our simulation with that of the commonly accepted ground truth for the same structure, i.e., a simulation performed using minimum coordinates. I will add comments to the corresponding files with a proper description. Would that be sufficient?
