+++ Tests/Pendulum_MinRealODE.m
@@ -2,6 +2,10 @@ function qDot = Pendulum_MinRealODE(t,q,tData)
 % /* This Source Code Form is subject to the terms of the Mozilla Public
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
 % * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+% 
+% ODE file for the minimum realization simulation of a single pendulum,
+% conventionally accepted as the ground truth in dynamics circles. 
+% 
 q1 = q(1);
 q2 = q(2);
 
