+++ Tests/TBar_MinRealODE.m
@@ -1,0 +1,32 @@
+function qDot = TBar_MinRealODE(t,q)
+% /* This Source Code Form is subject to the terms of the Mozilla Public
+% * License, v. 2.0. If a copy of the MPL was not distributed with this
+% * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+% 
+% ODE file for the minimum realization simulation of a TBar,
+% conventionally accepted as the ground truth in dynamics circles. 
+% 
+q1 = q(1);
+q2 = q(2);
+q3 = q(3);
+q4 = q(4);
+
+
+A = ...
+[ 1, 0,            0,            0;
+  0, 1,            0,            0;
+  0, 0, -(625*pi)/12,            0;
+  0, 0,            0, -(625*pi)/12];
+
+% With 0.7 rL K found using r and E
+b = ... 
+[                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   q3;
[…]
