### 2018-11-21 #1042 @vaishnavtv
@ctdegroot, as recommended I have added another test case to the repository. I'm amending the contributing guidelines to reflect this change.
