### 2018-11-22 #1042 @ctdegroot
@vaishnavtv Thanks for fixing these up. I've ticked the box for tests now, since there is enough to check the code to a reasonable degree.
