### 2018-11-22 #1042 @ctdegroot
My review is concluded. I'm satisfied with the changes.
