### 2018-11-22 #1042 @labarba
@vaishnavtv : I read over your revised paper and I only have one suggestion, and that is to delete the sentence: “The scientific computing community’s familiarity with MATLAB makes it a favored choice for performing simulations of this nature.”  (Unsubstantiated assessment.)

We now only await the recommendation to accept from @apsabelhaus.

Unofficially, I was surprised that you chose to deposit your preprint on ResearchGate, rather than arXiv, or [engrXiv](https://engrxiv.org). I encourage you to consider these options in the future.
