### 2018-11-23 e47d9c7
Deleted statement.
