+++ Paper/JOSSpaper.md
@@ -36,7 +36,7 @@ A tensegrity system is an arrangement of axially-loaded elements (no element ben
 
 The dynamics of tensegrity systems is governed by multi-body dynamics, given by a set of ordinary differential equations. We have developed a Lagrangian formulation for deriving these differential equations directly from the given topology of members (bars and strings), and their mass and geometric properties. Three key features of classical tensegrity systems are a) actuations only occur via cables, b) bar-to-bar connections are pin joints, and c) the bars do not spin about their longitudinal axis. These properties are exploited to simplify the equations of motion. However, the Lagrangian framework presented here is general enough to allow modeling of general multi-body systems with actuated joints.
 
-`STEDY` is a MATLAB package for conducting numerically accurate tensegrity dynamics. The scientific computing community's familiarity with MATLAB makes it a favored choice for performing simulations of this nature. The equations of motion are developed in Cartesian coordinates. Although the choice of non-minimum coordinates is fairly controversial in the robotics community on account of a reduction in the system's configuration space in the presence of joints, adopting Cartesian coordinates precludes any singularities developed in the mass matrix [@tenseBook] and facilitates derivations of elegant differential-algebraic equations governing the multi-body dynamics. However, usage of non-minimum coordinates is likely to result in the solution drifting away from the constraint space because of integration errors [@ConstraintViolation]. The software package includes an implementation of the direct correction scheme that minimizes numerically induced constraint violations and energy conservation errors. We have demonstrated superiority of the method in terms of accuracy over the commercially available dynamics simulator, Simscape Multibody [@matlab].
+`STEDY` is a MATLAB package for conducting numerically accurate tensegrity dynamics. The equations of motion are developed in non-minimum Cartesian coordinates. Although the choice of non-minimum coordinates is fairly controversial in the robotics community on account of a reduction in the system's configuration space in the presence of joints, adopting Cartesian coordinates precludes any singularities developed in the mass matrix [@tenseBook] and facilitates derivations of elegant differential-algebraic equations governing the multi-body dynamics. However, usage of non-minimum coordinates is likely to result in the solution drifting away from the constraint space because of integration errors [@ConstraintViolation]. The software package includes an implementation of the direct correction scheme that minimizes numerically induced constraint violations and energy conservation errors. We have demonstrated superiority of the method in terms of accuracy over the commercially available dynamics simulator, Simscape Multibody [@matlab].
 
 `STEDY` was designed to be used by researchers looking to simulate tensegrity dynamics without having to derive the differential equations for any structure in particular. The inputs that are required from the user are simply the parameters that uniquely define the tensegrity structure (initial nodal configuration, connectivity), the material and geometric properties, and simulation environment properties (inertially fixed nodes, external forces, duration, ODE solver options). Currently, `STEDY` is being used by TAMU researchers to develop [space habitats] and [wind turbine blades] using tensegrity principles.
 
