### 2018-11-23 #1042 @vaishnavtv
@labarba statement deleted.
Could you help explain to me why one should choose arXiv or any of the other OSF journals over ResearchGate to publish preprints? Is there a visibility advantage?
