+++ Docs/README.md
@@ -1,4 +1,5 @@
 [![Latest release](https://img.shields.io/github/release/uqLab/stedy.svg)](https://github.com/uqLab/stedy/releases)
+[![DOI](https://zenodo.org/badge/151340704.svg)](https://zenodo.org/badge/latestdoi/151340704)
 [![Documentation](https://img.shields.io/badge/documentation-user_guide-brightgreen.svg)](UserGuide.md)
 [![License: MPL 2.0](https://img.shields.io/badge/License-MPL%202.0-brightgreen.svg)](https://opensource.org/licenses/MPL-2.0)
 # Summary
