### 2018-12-26 #1000 @labarba
Wrong issue. We should be at https://github.com/openjournals/joss-reviews/issues/1042
