### 2018-12-30 #1042 @labarba
Hi there, @vaishnavtv — As soon as you correct the metadata of the Zenodo entry, I can get you paper published here in JOSS. Let me know!
